removed model_examination dir and moved its content
