+++ mindsdb/libs/phases/model_analyzer/helpers/column_evaluator.py
